pdf_converter/requirements.txt
# PDF Processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
pytesseract>=0.3.10

# LLM Integration
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.8.0

# Data Processing
pandas>=2.0.0
pydantic>=2.5.0

# Excel Generation
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0
